refactor: replace BuyButton logic to ProductCard
